requirements.txt
pyaudio
websockets
python-dotenv
fastapi
uvicorn
python-multipart
asyncio
dataclasses==0.8; python_version < "3.7"
typing-extensions==4.7.1; python_version < "3.8"